MVP/Cloud_Function/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud>=0.34.0
google-cloud-bigquery>=2.26.0
google-cloud-storage>=1.42.2
pandas>=1.3.3
shapely>=1.7.1
requests>=2.26.0
pyarrow>=5.0.0